fix: refactor - utils now
