+++ src/auto_apply_bot/model_interfaces/cover_letter_generator/generate_custom_model.py
@@ -1,7 +1,6 @@
 from auto_apply_bot.model_interfaces.cover_letter_generator.cover_letter_generator import CoverLetterModelInterface
 from auto_apply_bot import resolve_project_source
-from pathlib import Path
-from auto_apply_bot.logger import get_logger
+from auto_apply_bot.utils.logger import get_logger
 
 
 logger = get_logger(__name__)
