+++ src/auto_apply_bot/service/api/views.py
@@ -1,4 +1,4 @@
-from auto_apply_bot.logger import get_logger
+from auto_apply_bot.utils.logger import get_logger
 from django.http import JsonResponse, HttpRequest
 from django.conf import settings
 from django.views.decorators.csrf import csrf_exempt
